+++ os/mm/mm_heap/mm_getheap.c
@@ -20,6 +20,7 @@
  ****************************************************************************/
 
 #include <tinyara/config.h>
+#include <debug.h>
 #include <tinyara/mm/mm.h>
 
 extern struct mm_heap_s g_mmheap[CONFIG_MM_NHEAPS];
@@ -30,6 +31,45 @@ extern struct mm_heap_s g_mmheap[CONFIG_MM_NHEAPS];
 /****************************************************************************
  * Public Functions
  ****************************************************************************/
+/****************************************************************************
+ * Name: mm_get_heap
+ *
+ * Description:
+ *   returns a heap which type is matched with ttype
+ *
+ ****************************************************************************/
+struct mm_heap_s *mm_get_heap(void *address)
+{
+	int heap_idx;
+#ifdef CONFIG_MM_KERNEL_HEAP
+	struct tcb_s *tcb;
+
+	tcb = sched_gettcb(getpid());
+	if (tcb->flags & TCB_FLAG_TTYPE_MASK == TCB_FLAG_TTYPE_KERNEL) {
+		return &g_kmmheap;
+	} else
+#endif
+	{
+		heap_idx = mm_get_heapindex(address);
+		if (heap_idx == INVALID_HEAP_IDX) {
+			mdbg("address is not in heap region.\n");
+			return NULL;
+		}
+		return &g_mmheap[heap_idx];
+	}
+
+}
+/****************************************************************************
+ * Name: mm_get_heap_with_index
+ ****************************************************************************/
+struct mm_heap_s *mm_get_heap_with_index(int index)
+{
+	if (index >= CONFIG_MM_NHEAPS) {
+		mdbg("heap index is out of range.\n");
+		return NULL;
+	}
+	return &g_mmheap[index];
+}
 
 /****************************************************************************
  * Name: mm_get_heapindex
